(More-or-less complete) conversion of keys file to lua
keybindings are added by calling the addBinding method of default_keymode. It keys two
parameters: the keys to which the action is bound to and the action itself. The action can be
specified as a string (which is passed to CommandParser) or as a lua function.

Support for multiple keymodes is forthcoming. For the time being, I removed the BindKey and
KeyMode commands, as they will take a bit different form.