Don't create a new lua state upon reconfigure
As we integrate lua more and more into the core of fluxbox, it's going to progressively harder to
update all things that reference to the old state.
Furthermore, I would like to completely drop the reconfiguration at some point, after I'm
finished with making sure that all resources use the modifiedSig signal to update automatically.